+++ src/backend_dispatch.rs
@@ -1,6 +1,8 @@
 //! Implements `buffer_interface::*` traits for enums dispatching to backends
 
-use crate::{backend_interface::*, backends, AlphaMode, InitError, Pixel, Rect, SoftBufferError};
+use crate::{
+    backend_interface::*, backends, AlphaMode, InitError, Pixel, PixelFormat, Rect, SoftBufferError,
+};
 
 use raw_window_handle::{HasDisplayHandle, HasWindowHandle};
 use std::fmt;
@@ -100,29 +102,35 @@ macro_rules! make_dispatch {
             }
 
             #[inline]
-            fn supports_alpha_mode(&self, alpha_mode: AlphaMode) -> bool {
+            fn supported_pixel_formats(&self, alpha_mode: AlphaMode) -> &[PixelFormat] {
                 match self {
                     $(
                         $(#[$attr])*
-                        Self::$name(inner) => inner.supports_alpha_mode(alpha_mode),
+                        Self::$name(inner) => inner.supported_pixel_formats(alpha_mode),
                     )*
                 }
             }
 
-            fn configure(&mut self, width: NonZeroU32, height: NonZeroU32, alpha_mode: AlphaMode) -> Result<(), SoftBufferError> {
+            fn configure(
+                &mut self,
+                width: NonZeroU32,
+                height: NonZeroU32,
+                alpha_mode: AlphaMode,
+                pixel_format: PixelFormat,
+            ) -> Result<(), SoftBufferError> {
                 match self {
                     $(
                         $(#[$attr])*
-                        Self::$name(inner) => inner.configure(width, height, alpha_mode),
+                        Self::$name(inner) => inner.configure(width, height, alpha_mode, pixel_format),
                     )*
                 }
             }
 
-            fn next_buffer(&mut self, alpha_mode: AlphaMode) -> Result<BufferDispatch<'_>, SoftBufferError> {
+            fn next_buffer(&mut self, alpha_mode: AlphaMode, pixel_format: PixelFormat) -> Result<BufferDispatch<'_>, SoftBufferError> {
                 match self {
                     $(
                         $(#[$attr])*
-                        Self::$name(inner) => Ok(BufferDispatch::$name(inner.next_buffer(alpha_mode)?)),
+                        Self::$name(inner) => Ok(BufferDispatch::$name(inner.next_buffer(alpha_mode, pixel_format)?)),
                     )*
                 }
             }