+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Unreleased
 
 - Added `AlphaMode` to allow configuring transparency and zero-copy on Web. Set it with `Surface::configure`.
-- Added `Surface::supports_alpha_mode` for querying supported alpha modes.
-- Added `PixelFormat` enum.
+- Added `PixelFormat` enum. Set it with `Surface::configure`.
+- Added `Surface::supported_pixel_formats` for querying supported alpha modes and pixel formats.
 - Added `Buffer::pixels()` for accessing the buffer's pixel data.
 - Added `Buffer::pixel_rows()` for iterating over rows of the buffer data.
 - Added `Buffer::pixels_iter()` for iterating over each pixel with its associated `x`/`y` coordinate.